+++ src/ir_passes/mod.rs
@@ -1,1 +1,3 @@
 mod hoist_facts;
+
+pub use hoist_facts::HoistFacts;
