+++ src/ir_visitor/mod.rs
@@ -1,0 +1,3 @@
+mod visitor;
+
+pub use visitor::{Action, Visitor};
