+++ src/lib.rs
@@ -5,6 +5,7 @@ pub mod diagnostics;
 pub mod frontend;
 pub mod ir;
 pub mod ir_passes;
+pub mod ir_visitor;
 pub mod passes;
 pub mod resolver;
 pub mod utils;
