+++ src/main.rs
@@ -1,5 +1,6 @@
 use filament::{
-    backend, binding, cmdline, ir,
+    backend, binding, cmdline, ir, ir_passes,
+    ir_visitor::Visitor,
     passes::{self, Pass},
     resolver::Resolver,
     visitor::{Checker, Transform},
@@ -27,7 +28,8 @@ fn run(opts: &cmdline::Opts) -> Result<(), u64> {
     log::debug!("{ns}");
 
     if opts.show_ir {
-        let ir = ir::transform(ns);
+        let mut ir = ir::transform(ns);
+        ir_passes::HoistFacts::do_pass(&mut ir);
         ir::Printer::context(&ir, &mut std::io::stdout()).unwrap();
         return Ok(());
     }
